first results regarding the meaning of info/time

+++ src/MultiLevelHMCGenerator/MultiLevelHMCGenerator.h
@@ -78,7 +78,7 @@ public:
                          const std::string &name, HighFive::File &file);
 
     void analyze_dataset(const std::string &name, HighFive::File &file,
-                         size_t block_size, size_t size_to_use, size_t amount_of_sample_sets, size_t max_t);
+                         int block_size, int size_to_use, int amount_of_sample_sets, size_t max_t);
 
     /**
      * @brief Dumps the MultiLevelHMCGenerator, including all sub models and sub HMCGenerators
@@ -467,8 +467,8 @@ void MultiLevelHMCGenerator<configuration_type>::dump_observable(
 
 template<class configuration_type>
 void MultiLevelHMCGenerator<configuration_type>::analyze_dataset(const std::string &name, HighFive::File &file,
-                                                                 size_t block_size, size_t size_to_use,
-                                                                 size_t amount_of_sample_sets,
+                                                                 int block_size, int size_to_use,
+                                                                 int amount_of_sample_sets,
                                                                  size_t max_t) {
     for (int i = 0; i < HMCStack.size(); ++i) {
         std::string current_level_name{level_name};
